requirements.txt
# Streamlit
streamlit
pandas==1.3.5
numpy==1.23.5
wordcloud
matplotlib==3.5.3
Pillow==9.2.0

# Predict
tensorflow==2.8.3

nltk==3.7
gensim==4.2.0
scikit-learn==1.0.2

# Selenium
selenium==4.6.1
webdriver-manager==3.8.5
